refactor(card): Seperate Card and CardList component
